backend/requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.5.0
python-dotenv==1.0.0
groq==0.5.0
google-generativeai==0.3.2
numpy==2.3.3
tiktoken==0.5.2
pydantic==2.11.9
python-multipart
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-google-genai==2.1.12
langchain-text-splitters==0.3.11
google-ai-generativelanguage==0.4.0